+++ python/ry/JSON.pyi
@@ -6,3 +6,12 @@ from ry.ryo3.JSON import dumps as dumps
 from ry.ryo3.JSON import loads as loads
 from ry.ryo3.JSON import parse as parse
 from ry.ryo3.JSON import stringify as stringify
+
+__all__ = (
+    "cache_clear",
+    "cache_usage",
+    "dumps",
+    "loads",
+    "parse",
+    "stringify",
+)
