+++ python/ry/dirs.pyi
@@ -36,3 +36,44 @@ from ry.ryo3.dirs import template as template
 from ry.ryo3.dirs import template_dir as template_dir
 from ry.ryo3.dirs import video as video
 from ry.ryo3.dirs import video_dir as video_dir
+
+__all__ = (
+    "audio",
+    "audio_dir",
+    "cache",
+    "cache_dir",
+    "config",
+    "config_dir",
+    "config_local",
+    "config_local_dir",
+    "data",
+    "data_dir",
+    "data_local",
+    "data_local_dir",
+    "desktop",
+    "desktop_dir",
+    "document",
+    "document_dir",
+    "download",
+    "download_dir",
+    "executable",
+    "executable_dir",
+    "font",
+    "font_dir",
+    "home",
+    "home_dir",
+    "picture",
+    "picture_dir",
+    "preference",
+    "preference_dir",
+    "public",
+    "public_dir",
+    "runtime",
+    "runtime_dir",
+    "state",
+    "state_dir",
+    "template",
+    "template_dir",
+    "video",
+    "video_dir",
+)
