+++ python/ry/ryo3/xxhash.pyi
@@ -1,0 +1,98 @@
+import typing as t
+
+from ry._types import Buffer
+
+@t.final
+class Xxh32:
+    name: t.Literal["xxh32"]
+    digest_size: t.Literal[4]
+    block_size: t.Literal[16]
+
+    def __init__(self, input: Buffer = ..., seed: int | None = ...) -> None: ...
+    def update(self, input: Buffer) -> None: ...
+    def digest(self) -> bytes: ...
+    def hexdigest(self) -> str: ...
+    def intdigest(self) -> int: ...
+    def copy(self) -> Xxh32: ...
+    def reset(self, seed: int | None = ...) -> None: ...
+    @property
+    def seed(self) -> int: ...
+
+@t.final
+class Xxh64:
+    name: t.Literal["xxh64"]
+    digest_size: t.Literal[8]
+    block_size: t.Literal[32]
+
+    def __init__(self, input: Buffer | None = None, seed: int | None = ...) -> None: ...
+    def update(self, input: Buffer) -> None: ...
+    def digest(self) -> bytes: ...
+    def hexdigest(self) -> str: ...
+    def intdigest(self) -> int: ...
+    def copy(self) -> Xxh64: ...
+    def reset(self, seed: int | None = ...) -> None: ...
+    @property
+    def seed(self) -> int: ...
+
+@t.final
+class Xxh3:
+    name: t.Literal["xxh3"]
+    digest_size: int  # xxh3_64: 8, xxh3_128: 16
+    block_size: int  # xxh3_64: 32, xxh3_128: 64
+
+    def __init__(
+        self,
+        input: Buffer = ...,
+        seed: int | None = ...,
+        secret: bytes | None = ...,
+    ) -> None: ...
+    def update(self, input: Buffer) -> None: ...
+    def digest(self) -> bytes: ...
+    def hexdigest(self) -> str: ...
+    def intdigest(self) -> int: ...
+    @property
+    def seed(self) -> int: ...
+    def digest128(self) -> bytes: ...
+    def hexdigest128(self) -> str: ...
+    def intdigest128(self) -> int: ...
+    def copy(self) -> Xxh3: ...
+    def reset(self) -> None: ...
+
+# constructor aliases
+def xxh32(input: Buffer | None = None, seed: int | None = None) -> Xxh32: ...
+def xxh64(input: Buffer | None = None, seed: int | None = None) -> Xxh64: ...
+def xxh3(
+    input: Buffer | None = None, seed: int | None = None, secret: bytes | None = None
+) -> Xxh3: ...
+
+# -----------------------------------------------------------------------------
+# ONE-SHOT FUNCTIONS
+# -----------------------------------------------------------------------------
+
+# xxh32
+def xxh32_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh32_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
+def xxh32_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+
+# xxh64
+def xxh64_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh64_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
+def xxh64_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+
+# xxh128
+def xxh128_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh128_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
+def xxh128_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+
+# xxh3
+def xxh3_64_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh3_64_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+def xxh3_64_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
+def xxh3_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh3_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+def xxh3_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
+
+# xxh128
+def xxh3_128_digest(input: Buffer, seed: int | None = None) -> bytes: ...
+def xxh3_128_intdigest(input: Buffer, seed: int | None = None) -> int: ...
+def xxh3_128_hexdigest(input: Buffer, seed: int | None = None) -> str: ...
